backend/pyproject.toml
[project]
name = "Intervuo_backend"
version = "0.1.0"
description = "This is DJANGO APP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.1",
    "django-admin>=2.0.2",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt[blacklist]>=5.3.1",
    "djoser>=2.2.3",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.4.0",
    "mysqlclient>=2.2.4",
    "django-extensions>=3.2.3",
    "pillow>=10.4.0",
    "redis>=5.1.1",
    "celery>=5.4.0",
    "flower>=2.0.1",
    "langchain>=0.3.3",
    "langchain-community>=0.3.2",
    "langchain-openai>=0.2.2",
    "openai>=1.51.2",
    "scikit-learn>=1.5.2",
]
